Refactor logging extension
