+++ tests/app/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     "more_admin_filters",
     "polymorphic",
     "triggers",
+    "triggers.contrib.logging",
     "tests.app",
 ]
