+++ triggers/contrib/logging/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "triggers.contrib.logging.apps.LoggingConfig"
